Add oracle helper functions for price submission and fetching
