+++ payment-server/submitPrice.js
@@ -1,10 +1,17 @@
 import fs from "fs";
-import axios from "axios";
 import { assert } from "chai";
 import { rest, util } from "blockapps-rest";
 import http from "http";
 import BigNumber from "bignumber.js";
 
+import {
+  submitPrice,
+  updateAssetPrice,
+  fetchLBMAMetalPrice,
+  fetchERC20TokenPrice,
+  fetchMetalPrice,
+  fetchAsset,
+} from "./helpers/oracleHelper.js";
 import config from "./load.config.js";
 import deployment from "./load.deploy.js";
 import oauthHelper from "./helpers/oauthHelper.js";
@@ -62,53 +69,8 @@ if (process.env.SALE_UPDATE === "true") {
 // Global array to store all distributeRewards calls
 const distributeRewardsCallList = [];
 
-// Function to submit price to the oracle contract
-async function submitPrice(token, contract, args) {
-  const callArgs = {
-    contract,
-    method: "submitPrice",
-    args: util.usc(args),
-  };
-  await rest.call(token, callArgs, { config, cacheNonce: true });
-}
-
-// Function to update the price of the Asset Sale price
-async function updateAssetPrice(
-  assetMarkUp,
-  token,
-  contractAddress,
-  price,
-  decimals
-) {
-  const parsedPriceMarkup = parseFloat(assetMarkUp) || 1;
-
-  // Use BigNumber for precise calculation
-  const priceWithMarkup = new BigNumber(price)
-    .times(parsedPriceMarkup);
-
-  // Round to 2 decimal places using decimalPlaces with ROUND_HALF_UP
-  const priceBig = priceWithMarkup
-    .decimalPlaces(2, BigNumber.ROUND_HALF_UP)
-    .shiftedBy(-decimals)
-    .toNumber();
-
-  const callArgs = {
-    contract: {
-      address: contractAddress,
-    },
-    method: "update",
-    args: {
-      _quantity: 0,
-      _price: priceBig,
-      _paymentServices: [{ creator: "", serviceName: "" }],
-      _scheme: 2,
-    },
-  };
-  await rest.call(token, callArgs, { config, cacheNonce: true });
-}
-
 // Instead of calling rest.call directly, accumulate call arguments for distributeRewards
-async function distributeRewards(token, contract, args) {
+async function distributeRewards(contract, args) {
   const callArgs = {
     contract,
     method: "distributeRewards",
@@ -218,7 +180,6 @@ async function fetchAndSubmitEscrowAddresses(oracleContract, token) {
       console.log(`No escrows found for reserve ${reserveName}`);
       // Collect callArgs for empty escrowAddresses to distributeRewards
       await distributeRewards(
-        token,
         { address: reserveAddress, name: reserveName },
         { escrowAddresses: [] }
       );
@@ -238,7 +199,6 @@ async function fetchAndSubmitEscrowAddresses(oracleContract, token) {
 
       // Collect callArgs instead of calling directly
       await distributeRewards(
-        token,
         { address: reserveAddress, name: reserveName },
         { escrowAddresses }
       );
@@ -248,213 +208,6 @@ async function fetchAndSubmitEscrowAddresses(oracleContract, token) {
       `Escrow Addresses submitted for ${reserveName} at ${new Date().toISOString()}`
     );
   }
-}
-
-// Function to fetch and submit price
-async function fetchMetalPrice(metal, apiKey) {
-  try {
-    const apiUrl = `https://api.metals.dev/v1/metal/spot?metal=${metal}&api_key=${apiKey}&currency=USD&unit=toz`;
-    const response = await axios.get(apiUrl);
-    const metalPrice = response.data.rate.price;
-    console.log(`Current ${metal} Price: $${metalPrice} per ounce`);
-
-    const timestampInSeconds = Math.floor(Date.now() / 1000);
-    console.log(`Current Timestamp: ${timestampInSeconds}`);
-
-    return { price: metalPrice, timestampInSeconds };
-  } catch (error) {
-    console.error(`ERROR: Failed to fetch price for ${metal}:`, error);
-    await flagFile.appendToErrorFile(
-      `Failed to fetch price for ${metal}: ${error}`
-    );
-  }
-}
-
-// Function to fetch  and submit price
-async function fetchLBMAMetalPrice(metal, apiKey) {
-  try {
-    const apiUrl = `https://api.metals.dev/v1/metal/authority?api_key=${apiKey}&authority=lbma&currency=USD&unit=toz`;
-    const response = await axios.get(apiUrl);
-    const rates = response.data.rates;
-    let metalPrice;
-
-    if (metal.toLowerCase().includes("gold")) {
-      metalPrice = rates.lbma_gold_am;
-    } else if (metal.toLowerCase().includes("silver")) {
-      metalPrice = rates.lbma_silver;
-    } else {
-      throw new Error(`Metal ${metal} not supported`);
-    }
-
-    console.log(`Current ${metal} Price: $${metalPrice} per ounce`);
-
-    const timestampInSeconds = Math.floor(Date.now() / 1000);
-    console.log(`Current Timestamp: ${timestampInSeconds}`);
-
-    return { price: metalPrice, timestampInSeconds };
-  } catch (error) {
-    console.error(`ERROR: Failed to fetch price for ${metal}:`, error);
-    await flagFile.appendToErrorFile(
-      `Failed to fetch price for ${metal}: ${error}`
-    );
-  }
-}
-
-// Function to fetch and submit ETH price
-async function fetchERC20TokenPrice(
-  name,
-  apiKey
-) {
-  try {
-    const apiUrl = `https://api.g.alchemy.com/prices/v1/${apiKey}/tokens/historical`;
-    const currentTimeMs = Date.now();
-    const OneDayHoursInMs = 24 * 60 * 60 * 1000;
-    const requestBody = {
-      symbol: name,
-      startTime: Math.floor((currentTimeMs - OneDayHoursInMs) / 1000),
-      endTime: Math.floor(currentTimeMs / 1000),
-      interval: "1h",
-    };
-    const response = await axios.post(apiUrl, requestBody, {
-      headers: { "Content-Type": "application/json" },
-    });
-    const responseData = response.data;
-    if (!responseData?.data || !Array.isArray(responseData.data)) {
-      console.error("Invalid response format:", responseData);
-      throw new Error("Invalid price data format from API");
-    }
-    console.log(`Received ${responseData.data.length} price points`);
-    const prices = responseData.data.map(({ value, timestamp }) => ({
-      price: parseFloat(value),
-      timestamp: new Date(timestamp).getTime() / 1000,
-    }));
-    prices.forEach((point, index) => {
-      if (!Number.isFinite(point.price) || point.price <= 0) {
-        throw new Error(`Invalid price at index ${index}: ${point.price}`);
-      }
-      if (!Number.isFinite(point.timestamp) || point.timestamp <= 0) {
-        throw new Error(
-          `Invalid timestamp at index ${index}: ${point.timestamp}`
-        );
-      }
-    });
-    const twap = calculateTWAP(prices);
-    console.log(`Calculated TWAP: $${twap}`);
-    const currentTimestamp = Math.floor(currentTimeMs / 1000);
-    return {price: twap, timestamp: currentTimestamp}
-  }
-  catch (error) {
-    console.error("ETH TWAP calculation and submission failed:", error);
-    await flagFile.appendToErrorFile(
-      `ETH TWAP calculation and submission failed: ${error}`
-    );
-    throw error;
-  }
-}
-
-// Function to fetch and submit ETH price
-async function fetchAndSubmitERC20TokenPrice(
-  name,
-  oracleAddress,
-  decimals,
-  apiKey,
-  token
-) {
-  try {
-    const apiUrl = `https://api.g.alchemy.com/prices/v1/${apiKey}/tokens/historical`;
-
-    // Current time in milliseconds
-    const currentTimeMs = Date.now();
-    const OneDayHoursInMs = 24 * 60 * 60 * 1000;
-
-    // Define the request body
-    const requestBody = {
-      symbol: name,
-      startTime: Math.floor((currentTimeMs - OneDayHoursInMs) / 1000),
-      endTime: Math.floor(currentTimeMs / 1000),
-      interval: "1h",
-    };
-
-    // Make the POST request with the body
-    const response = await axios.post(apiUrl, requestBody, {
-      headers: { "Content-Type": "application/json" },
-    });
-
-    const responseData = response.data;
-
-    if (!responseData?.data || !Array.isArray(responseData.data)) {
-      console.error("Invalid response format:", responseData);
-      throw new Error("Invalid price data format from API");
-    }
-
-    console.log(`Received ${responseData.data.length} price points`);
-
-    const prices = responseData.data.map(({ value, timestamp }) => ({
-      price: parseFloat(value),
-      timestamp: new Date(timestamp).getTime() / 1000,
-    }));
-
-    prices.forEach((point, index) => {
-      if (!Number.isFinite(point.price) || point.price <= 0) {
-        throw new Error(`Invalid price at index ${index}: ${point.price}`);
-      }
-      if (!Number.isFinite(point.timestamp) || point.timestamp <= 0) {
-        throw new Error(
-          `Invalid timestamp at index ${index}: ${point.timestamp}`
-        );
-      }
-    });
-
-    const twap = calculateTWAP(prices);
-    console.log(`Calculated TWAP: $${twap}`);
-
-    const currentTimestamp = Math.floor(currentTimeMs / 1000);
-    // Round TWAP to 2 decimal places and adjust for token decimals
-    const priceBig = new BigNumber(twap)
-      .decimalPlaces(2, BigNumber.ROUND_HALF_UP)
-      .shiftedBy(-decimals)
-      .toNumber();
-
-    await submitPrice(
-      token,
-      { address: oracleAddress },
-      {
-        price: priceBig,
-        timestamp: currentTimestamp,
-      }
-    );
-
-    console.log(
-      `TWAP submitted: $${twap.toFixed(2)} at ${new Date(
-        currentTimeMs
-      ).toISOString()}`
-    );
-  } catch (error) {
-    console.error("ETH TWAP calculation and submission failed:", error);
-    await flagFile.appendToErrorFile(
-      `ETH TWAP calculation and submission failed: ${error}`
-    );
-    throw error;
-  }
-}
-
-function calculateTWAP(prices) {
-  let totalWeightedPrice = 0;
-  let totalWeight = 0;
-
-  for (let i = 1; i < prices.length; i++) {
-    const price = prices[i].price;
-    const timeDiff = prices[i].timestamp - prices[i - 1].timestamp;
-
-    totalWeightedPrice += price * timeDiff;
-    totalWeight += timeDiff;
-  }
-
-  if (totalWeight === 0) {
-    throw new Error("Invalid data for TWAP calculation");
-  }
-
-  return totalWeightedPrice / totalWeight;
 }
 
 // Function to submit oracle prices periodically
@@ -473,31 +226,40 @@ const submitOraclePricePeriodically = async () => {
       console.warn(`[Oracle WARN] Skipping invalid oracle ${key}`);
       continue;
     }
-
+    let priceBig = 0;
+    let result = null;
     try {
       if (oracle.type === "ERC20") {
-        await fetchAndSubmitERC20TokenPrice(
+        result = await fetchERC20TokenPrice(
           oracle.name,
           oracle.address,
           oracle.decimals,
           process.env.ALCHEMY_API_KEY,
           token
         );
-      } else if (oracle.type === "Metal") {
-        const metalResult = await fetchMetalPrice(
-          oracle.name.toLowerCase().replace(/st$/, ""),
-          process.env.METALS_API_KEY
-        );
-        if (metalResult) {
+        if (result) {
           // Round price to 2 decimal places and adjust for token decimals
-          const priceBig = new BigNumber(metalResult.price)
+          priceBig = new BigNumber(result.price)
             .decimalPlaces(2, BigNumber.ROUND_HALF_UP)
             .shiftedBy(-oracle.decimals)
             .toNumber();
-          await submitPrice(token, oracle, {
-            price: priceBig,
-            timestamp: metalResult.timestampInSeconds,
-          });
+          console.log(
+            `[Oracle Update] Price submitted for ${
+              oracle.name
+            } at ${new Date().toISOString()}`
+          );
+        }
+      } else if (oracle.type === "Metal") {
+        result = await fetchMetalPrice(
+          oracle.name.toLowerCase().replace(/st$/, ""),
+          process.env.METALS_API_KEY
+        );
+        if (result) {
+          // Round price to 2 decimal places and adjust for token decimals
+          priceBig = new BigNumber(result.price)
+            .decimalPlaces(2, BigNumber.ROUND_HALF_UP)
+            .shiftedBy(-oracle.decimals)
+            .toNumber();
           console.log(
             `[Oracle Update] Price submitted for ${
               oracle.name
@@ -505,18 +267,22 @@ const submitOraclePricePeriodically = async () => {
           );
         }
       } else if (oracle.type === "Constant") {
-        const priceBig = new BigNumber(oracle.price)
+        priceBig = new BigNumber(oracle.price)
           .decimalPlaces(2, BigNumber.ROUND_HALF_UP)
           .shiftedBy(-oracle.decimals)
           .toNumber();
-        await submitPrice(token, oracle, {
-          price: priceBig,
-          timestamp: Math.floor(Date.now() / 1000),
-        });
       } else {
         console.warn(`[Oracle WARN] Skipping unsupported oracle type ${key}`);
       }
-
+      await submitPrice(
+        token,
+        oracle,
+        {
+          price: priceBig,
+          timestamp: Math.floor(Date.now() / 1000),
+        },
+        config
+      );
       await fetchAndSubmitEscrowAddresses(oracle, token);
     } catch (error) {
       console.error(`[Oracle ERROR] Failed to process oracle ${key}:`, error);
@@ -538,20 +304,25 @@ const updateSalePricePeriodically = async () => {
   let result;
   for (const asset of assets) {
     try {
+      // get the correct asset with sale
+      
       const searchOptions = {
         config,
         query: {
-          address: "eq." + asset.address,
+          root: "eq." + asset.address,
           select: "sale,name,decimals",
         },
       };
 
-      const assetResult = await rest.search(
+      const assetResult = await fetchAsset(
         token,
-        { name: "BlockApps-Mercata-Asset" },
+        { address: asset.address },
         searchOptions
       );
-
+      if (!assetResult || assetResult.length === 0) {
+        console.warn(`[Sale Update] No asset found for ${asset.address}`);
+        continue;
+      }
       if (!assetResult[0]?.sale) {
         console.warn(`[Sale Update] Skipping invalid asset ${asset.address}`);
         continue;
@@ -569,14 +340,19 @@ const updateSalePricePeriodically = async () => {
         );
       }
 
-      const decimals = assetResult[0].decimals || 0;
+      const decimals = assetResult[0].decimals
+        ? assetResult[0].decimals
+        : assetResult[0].name.toLowerCase().includes("eth")
+        ? 18
+        : 0;
 
       await updateAssetPrice(
         asset.markUp,
         token,
         assetResult[0]?.sale,
         result.price,
-        decimals
+        decimals,
+        config
       );
       console.log(
         `[Sale Update] Price updated for asset: ${
